Statusbar progress widget: show button immediately
Summary:
This fixes the bug where the progress dialog shows up, and the user has
no button to close it down.

This happened because for very quick jobs, by the time the delay timer kicked in,
the list of progress items is actually empty already. But the progress
dialog stays up a bit longer...

Test Plan: Sync, abort, delete one mail in kmail

Reviewers: mlaurent, dvratil

Reviewed By: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D21065