Fix metainfo.yaml
The cmakename key is a single top-level entry; it doesn't belong
in the middle of the list of libraries. A repo can have only
one single cmakename, even if it has multiple libraries
(which is a list-valued entry).

This same problem (multiple cmakename keys, often in the middle
of the libraries list) occurs in multiple KDE-PIM repositories.