Avoid spurious checkcombobox toggle where indexes don't match
When clicking on a KCheckComboBox that has an item already selected,
QComboBox will fire activated (via _q_editingFinished). When this
popup is opening for the first time the currentIndex of the combo
box is correct but the view's current index is still row 0. This
leads to spuriously toggling the first item in the list.

Since this activation from QComboBox happens programatically, care
must also be taken not to process this event for a disabled item.

BUG: 347045